refactor(client): migrate antigravity/gemini from hook to xxx_client pattern
The hook pattern (oauthHookFunctions map + per-issuer switches inside
createSessionBoundTransport / the legacy CreateHTTPClientForProvider) had
provider wire-format details scattered across http.go. Aligns Gemini and
Antigravity with the existing Claude/Codex pattern: each provider's request
envelope, headers, and response unwrapping live in a dedicated *_client.go.

- gemini_client.go: NewGeminiClient builds a *GoogleClient whose http.Client
  uses geminiRoundTripper (URL rewrite to /v1internal:<op>, body wrap with
  {model, project, user_prompt_id, request}, Bearer auth, response unwrap).
- antigravity_client.go: same shape with the desktop-Antigravity envelope
  (project/requestId/userAgent/requestType) and antigravity User-Agent.
- code_assist_envelope.go: shared streaming/JSON unwrap helpers and the
  isStreamingRequest predicate, used by both clients.
- google.go: factor out newGoogleClientFromHTTPClient so the dedicated
  clients can inject their own http.Client without going through the
  generic transport path.
- pool.go: GetGoogleClient now dispatches via newGoogleClientForProvider —
  Gemini and Antigravity providers get their xxx_client constructors, and
  the embedded *GoogleClient is returned so forwarding code is unchanged.
- http.go: drop HookFunc, oauthHookFunctions, antigravityHook,
  requestModifier, both round-tripper types, and the dead-code
  CreateHTTPClientForProvider. createSessionBoundTransport is now
  generic — provider switches removed.

Tests:
- TestCreateSessionBoundTransport_{Antigravity,Gemini,Codex} updated to
  assert the transport layer stays generic.
- New TestNewGeminiClient_* and TestNewAntigravityClient_* verify the
  xxx_client constructors apply the right round tripper.
- Existing geminiRoundTripper envelope tests continue to pass against the
  moved type.
